refactor: config 리팩토링
