+++ src/main/java/gift/config/MethodOverrideConfig.java
@@ -1,0 +1,14 @@
+package gift.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.filter.HiddenHttpMethodFilter;
+
+@Configuration
+public class MethodOverrideConfig {
+
+    @Bean
+    public HiddenHttpMethodFilter hiddenHttpMethodFilter() {
+        return new HiddenHttpMethodFilter();
+    }
+}
